split Workouts into WorkoutList and Workout
